requirements.txt
# Pytest test
pytest==6.2.4

# Pytest plugins
pytest-spec==3.2.0
pytest-html-reporter==0.2.6
pytest-html==3.1.1
pytest-ordering==0.6
coverage==4.3.4
pytest-cov==2.4.0
pytest-xdist==2.5.0

# HTTP library
requests==2.25.1

# JSON encoder and decoder
simplejson==3.17.2

# Webdriver manager
webdriver-manager==3.4.2

# Selenium
selenium==3.141.0

# Yaml
PyYAML==5.4.1

# JWT token decode
# PyJWT==1.7.1

# URI template to extend URL
uritemplate==3.0.1

# Retry functionality with decorators
# retry==0.9.2

# Sqlalchemy
SQLAlchemy==1.4.32

# postgressql
# if postgres is not working, try with anyone below command and install postgresql
# 'sudo apt-get install build-dep python-psycopg2' (or)
# 'sudo apt-get install libpq-dev'
# Now install postgres sql
# 'pip install psycopg2' to install postgressql dependency
# psycopg2==2.9.3

# dotenv
python-dotenv==0.20.0

# python-docx
python-docx==0.8.11